+++ BytesToBits.m
@@ -3,5 +3,5 @@ function bits = BytesToBits(bytes)
     % 输入: bytes - uint8数组
     % 输出: bits - uint8数组，值为0或1
     bits = bitget(bytes(:), 8:-1:1)';
-    bits = bits(:)';
+    bits =uint8( bits(:)');
 end
